+++ scheduler/heartbeat.go
@@ -79,7 +79,7 @@ func (e *Scheduler) Heartbeat() {
 			task.Discovery = e.changeDiscoveryInfo(task)
 
 			// add task to communication channel
-			e.Framework.CommandChan <- task
+			e.Framework.CommandChan <- *task
 
 			e.Redis.SaveTaskRedis(task)