+++ scheduler/scheduler.go
@@ -2,7 +2,10 @@ package scheduler
 
 import (
 	"bufio"
+	"bytes"
+	"crypto/tls"
 	"net/http"
+	"strconv"
 	"strings"
 
 	api "github.com/AVENTER-UG/mesos-compose/api"
@@ -12,8 +15,7 @@ import (
 	cfg "github.com/AVENTER-UG/mesos-compose/types"
 	"github.com/AVENTER-UG/util/vault"
 	"github.com/sirupsen/logrus"
-
-	"github.com/gogo/protobuf/jsonpb"
+	"google.golang.org/protobuf/encoding/protojson"
 )
 
 // Scheduler include all the current vars and global config
@@ -29,15 +31,47 @@ type Scheduler struct {
 	ConnectionError bool
 }
 
-// New will create a new Scheduler object
-func New(cfg *cfg.Config, frm *cfg.FrameworkConfig) *Scheduler {
+// Marshaler to serialize Protobuf Message to JSON
+var marshaller = protojson.MarshalOptions{
+	UseEnumNumbers: false,
+	Indent:         " ",
+	UseProtoNames:  true,
+}
+
+// Subscribe to the mesos backend
+func Subscribe(cfg *cfg.Config, frm *cfg.FrameworkConfig) *Scheduler {
 	e := &Scheduler{
 		Config:    cfg,
 		Framework: frm,
 		Mesos:     *mesos.New(cfg, frm),
 	}
 
-	e.Client, e.Req = e.Mesos.Subscribe()
+	subscribeCall := &mesosproto.Call{
+		FrameworkId: e.Framework.FrameworkInfo.Id,
+		Type:        mesosproto.Call_SUBSCRIBE.Enum(),
+		Subscribe: &mesosproto.Call_Subscribe{
+			FrameworkInfo: &e.Framework.FrameworkInfo,
+		},
+	}
+	logrus.WithField("func", "scheduler.Subscribe").Debug(subscribeCall)
+	body, _ := marshaller.Marshal(subscribeCall)
+	client := &http.Client{}
+	// #nosec G402
+	client.Transport = &http.Transport{
+		TLSClientConfig: &tls.Config{InsecureSkipVerify: e.Config.SkipSSL},
+	}
+
+	protocol := "https"
+	if !e.Framework.MesosSSL {
+		protocol = "http"
+	}
+	req, _ := http.NewRequest("POST", protocol+"://"+e.Framework.MesosMasterServer+"/api/v1/scheduler", bytes.NewBuffer([]byte(body)))
+	req.Close = true
+	req.SetBasicAuth(e.Framework.Username, e.Framework.Password)
+	req.Header.Set("Content-Type", "application/json")
+
+	e.Req = req
+	e.Client = client
 
 	return e
 }
@@ -55,7 +89,7 @@ func (e *Scheduler) EventLoop() {
 	reader := bufio.NewReader(res.Body)
 
 	line, _ := reader.ReadString('\n')
-	_ = strings.TrimSuffix(line, "\n")
+	bytesCount, _ := strconv.Atoi(strings.Trim(line, "\n"))
 
 	go e.HeartbeatLoop()
 	go e.ReconcileLoop()
@@ -63,34 +97,46 @@ func (e *Scheduler) EventLoop() {
 	for {
 		// Read line from Mesos
 		line, err = reader.ReadString('\n')
+		_ = strings.Trim(line, "\n")
 		if err != nil {
 			logrus.WithField("func", "scheduler.EventLoop").Error("Error to read data from Mesos Master: ", err.Error())
 			return
 		}
-		line = strings.TrimSuffix(line, "\n")
+
+		// skip if no data
+		if line == "" || len(line)-1 < bytesCount {
+			logrus.WithField("func", "scheduler.EventLoop").Tracef("Line %s, bytesCount: %d ", line, bytesCount)
+			logrus.WithField("func", "scheduler.EventLoop").Trace("No data from Mesos Master")
+			continue
+		}
+		data := line[:bytesCount]
+		bytesCount, _ = strconv.Atoi(line[bytesCount : len(line)-1])
+
 		// Read important data
 		var event mesosproto.Event // Event as ProtoBuf
-		err := jsonpb.UnmarshalString(line, &event)
+		err := protojson.Unmarshal([]byte(data), &event)
 		if err != nil {
 			logrus.WithField("func", "scheduler.EventLoop").Warn("Could not unmarshal Mesos Master data: ", err.Error())
 			continue
 		}
 
-		switch event.Type {
-		case mesosproto.Event_SUBSCRIBED:
+		logrus.WithField("func", "scheduler.EventLoop").Tracef("Event %s", event.GetType().String())
+
+		switch event.Type.Number() {
+		case mesosproto.Event_SUBSCRIBED.Number():
 			logrus.WithField("func", "scheduler.EventLoop").Info("Subscribed")
-			logrus.WithField("func", "scheduler.EventLoop").Debug("FrameworkId: ", event.Subscribed.GetFrameworkID())
-			e.Framework.FrameworkInfo.ID = event.Subscribed.GetFrameworkID()
+			logrus.WithField("func", "scheduler.EventLoop").Debug("FrameworkId: ", event.Subscribed.GetFrameworkId())
+			e.Framework.FrameworkInfo.Id = event.Subscribed.GetFrameworkId()
 			e.Framework.MesosStreamID = res.Header.Get("Mesos-Stream-Id")
 
 			go e.reconcile()
-			go e.Redis.SaveFrameworkRedis(*e.Framework)
+			go e.Redis.SaveFrameworkRedis(e.Framework)
 			go e.Redis.SaveConfig(*e.Config)
-		case mesosproto.Event_UPDATE:
+		case mesosproto.Event_UPDATE.Number():
 			go e.HandleUpdate(&event)
 			go e.Redis.SaveConfig(*e.Config)
-			go e.callPluginEvent(event)
-		case mesosproto.Event_OFFERS:
+			go e.callPluginEvent(&event)
+		case mesosproto.Event_OFFERS.Number():
 			// Search Failed containers and restart them
 			err = e.HandleOffers(event.Offers)
 			if err != nil {
@@ -100,8 +146,8 @@ func (e *Scheduler) EventLoop() {
 	}
 }
 
-func (e *Scheduler) changeDockerPorts(cmd cfg.Command) []mesosproto.ContainerInfo_DockerInfo_PortMapping {
-	var ret []mesosproto.ContainerInfo_DockerInfo_PortMapping
+func (e *Scheduler) changeDockerPorts(cmd *cfg.Command) []*mesosproto.ContainerInfo_DockerInfo_PortMapping {
+	var ret []*mesosproto.ContainerInfo_DockerInfo_PortMapping
 	for _, port := range cmd.DockerPortMappings {
 		port.HostPort = e.API.GetRandomHostPort()
 		ret = append(ret, port)
@@ -109,16 +155,16 @@ func (e *Scheduler) changeDockerPorts(cmd cfg.Command) []mesosproto.ContainerInf
 	return ret
 }
 
-func (e *Scheduler) changeDiscoveryInfo(cmd cfg.Command) mesosproto.DiscoveryInfo {
+func (e *Scheduler) changeDiscoveryInfo(cmd *cfg.Command) *mesosproto.DiscoveryInfo {
 	for i, port := range cmd.DockerPortMappings {
 		cmd.Discovery.Ports.Ports[i].Number = port.HostPort
 	}
 	return cmd.Discovery
 }
 
-// Reconcile will reconcile the task states after the framework was restarted
+// reconcile will ask Mesos about the current state of the given tasks
 func (e *Scheduler) reconcile() {
-	var oldTasks []mesosproto.Call_Reconcile_Task
+	var oldTasks []*mesosproto.Call_Reconcile_Task
 	keys := e.Redis.GetAllRedisKeys(e.Framework.FrameworkName + ":*")
 	for keys.Next(e.Redis.CTX) {
 		// continue if the key is not a mesos task
@@ -134,31 +180,31 @@ func (e *Scheduler) reconcile() {
 			continue
 		}
 
-		oldTasks = append(oldTasks, mesosproto.Call_Reconcile_Task{
-			TaskID: mesosproto.TaskID{
-				Value: task.TaskID,
+		oldTasks = append(oldTasks, &mesosproto.Call_Reconcile_Task{
+			TaskId: &mesosproto.TaskID{
+				Value: &task.TaskID,
 			},
-			AgentID: &mesosproto.AgentID{
-				Value: task.Agent,
+			AgentId: &mesosproto.AgentID{
+				Value: &task.MesosAgent.ID,
 			},
 		})
 		logrus.WithField("func", "mesos.Reconcile").Debug("Reconcile Task: ", task.TaskID)
 	}
 	err := e.Mesos.Call(&mesosproto.Call{
-		Type:      mesosproto.Call_RECONCILE,
+		Type:      mesosproto.Call_RECONCILE.Enum(),
 		Reconcile: &mesosproto.Call_Reconcile{Tasks: oldTasks},
 	})
 
 	if err != nil {
-		logrus.WithField("func", "mesos.Reconcile").Debug("Reconcile Error: ", err)
+		logrus.WithField("func", "scheduler.reconcile").Debug("Reconcile Error: ", err)
 	}
 }
 
 // implicitReconcile will ask Mesos which tasks and there state are registert to this framework
 func (e *Scheduler) implicitReconcile() {
-	var noTasks []mesosproto.Call_Reconcile_Task
+	var noTasks []*mesosproto.Call_Reconcile_Task
 	err := e.Mesos.Call(&mesosproto.Call{
-		Type:      mesosproto.Call_RECONCILE,
+		Type:      mesosproto.Call_RECONCILE.Enum(),
 		Reconcile: &mesosproto.Call_Reconcile{Tasks: noTasks},
 	})
 
@@ -167,7 +213,7 @@ func (e *Scheduler) implicitReconcile() {
 	}
 }
 
-func (e *Scheduler) callPluginEvent(event mesosproto.Event) {
+func (e *Scheduler) callPluginEvent(event *mesosproto.Event) {
 	if e.Config.PluginsEnable {
 		for _, p := range e.Config.Plugins {
 			symbol, err := p.Lookup("Event")
@@ -176,7 +222,7 @@ func (e *Scheduler) callPluginEvent(event mesosproto.Event) {
 				continue
 			}
 
-			eventPluginFunc, ok := symbol.(func(mesosproto.Event))
+			eventPluginFunc, ok := symbol.(func(*mesosproto.Event))
 			if !ok {
 				logrus.WithField("func", "main.initPlugins").Error("Error plugin does not have init function")
 				continue